move revision browser to use redux
