+++ app/addons/documents/rev-browser/actiontypes.js
@@ -10,9 +10,14 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
+export const REV_BROWSER_REV_TREE_LOADED = 'REV_TREE_LOADED';
+export const REV_BROWSER_DIFF_DOCS_READY = 'REV_BROWSER_DIFF_DOCS_READY';
+export const REV_BROWSER_DIFF_ENABLE_DIFF_VIEW = 'REV_BROWSER_DIFF_ENABLE_DIFF_VIEW';
+export const REV_BROWSER_SHOW_CONFIRM_MODAL = 'REV_BROWSER_SHOW_CONFIRM_MODAL';
+
 export default {
-  REV_BROWSER_REV_TREE_LOADED: 'REV_TREE_LOADED',
-  REV_BROWSER_DIFF_DOCS_READY: 'REV_BROWSER_DIFF_DOCS_READY',
-  REV_BROWSER_DIFF_ENABLE_DIFF_VIEW: 'REV_BROWSER_DIFF_ENABLE_DIFF_VIEW',
-  REV_BROWSER_SHOW_CONFIRM_MODAL: 'REV_BROWSER_SHOW_CONFIRM_MODAL'
+  REV_BROWSER_REV_TREE_LOADED,
+  REV_BROWSER_DIFF_DOCS_READY,
+  REV_BROWSER_DIFF_ENABLE_DIFF_VIEW,
+  REV_BROWSER_SHOW_CONFIRM_MODAL
 };
